+++ packages/components/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 7.9.2
+
+### Patch Changes
+
+- Fix docs
+- Updated dependencies [undefined]
+  - @heathmont/moon-assets@7.9.2
+  - @heathmont/moon-themes@7.9.2
+  - @heathmont/moon-utils@7.9.2
+
 ## 7.9.1
 
 ### Patch Changes